% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdbs.R
\docType{package}
\name{cfdbs}
\alias{cfdbs}
\alias{cfdbs-package}
\title{cfdbs: Extract info from Database}
\description{
Tools to aid in extraction of data from the cfdbs database and ADIOS!
}
\details{
\itemize{
\item Get_ functions to access, species coes, gear codes, landings etc.
\item Bundled data aggregating area codes to regions and gear codes to fleets
\item Secure login to servers
}

To learn more about using \code{cfdbs}, start with the vignette: \code{browseVignettes(package="cfdbs")} or click the index below
}
