% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTonneageString.R
\name{createTonnageString}
\alias{createTonnageString}
\title{Utility function (Internal, not exported) to convert argument to strings}
\usage{
createTonnageString(itemName, chosenItem)
}
\arguments{
\item{itemName}{Character string. Variable name as it exists in the data base}

\item{chosenItem}{User input value of variable listed in itemName}
}
\value{
A charachter string
}
\description{
inputs to get_ functions are required as strings when passed as a sql statement
This function converts numeric inputs to character strings. Used specifically for tonnage class
}
\examples{
\dontrun{
createTonnageString(itemName="toncl2",tonnage=c(1,2))
createTonnageString(itemName="toncl2",tonnage="all"))

}
}
