% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPUs.R
\docType{data}
\name{EPUs}
\alias{EPUs}
\title{EPU designations by statistical areas}
\format{A list containing 4 elements
\describe{
    \item{codes}{a vector of unique statistical areas}
    \item{names}{a character vector of EPU names}
    \item{description}{a character vector describing the EPU name}
    \item{data}{a list of 4 EPUs with associated statistical areas}
}}
\source{
The EPU/statistical area associations were a collaborative effort among staff at the ecosystem dynamics and assessment branch (EDAB) at the NEFSC
}
\usage{
EPUs
}
\description{
A list of 4 Ecological Production Units (EPUs) and the statistical areas that comprise each area
}
\seealso{
For a complete list of all statistical areas see: \code{\link{get_areas}}
}
\keyword{datasets}
