% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_length.R
\name{get_age_length}
\alias{get_age_length}
\title{Extract Length at Age data CFDBS}
\usage{
get_age_length(channel, year = 1994, species = "all",
  species_itis = FALSE, sex = "all")
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{species}{A specific species code or set of codes. Either numeric or character vector. Defaults to "all" species.
Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement. Character codes are short character strings.}

\item{sex}{Character vector. Default = "all". options "M" (male), "F" (female), "U" (unsexed)}
}
\value{
A list is returned:

  \item{data}{containing the result of the executed \code{$sql} statement}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}

The default sql statement "\code{select * from stockeff.mv_cf_age}" is used
}
\description{
Extract a list of length and ages for speices sampled in Commercial fisheries data.
This data is extracted from ADIOS (MV_CF_AGE)
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations. 
Note: species codes (nespp3) are stored in the database as VARCHAR2(3 BYTE)
}

\examples{
\dontrun{
# extracts all length and age data for all species defined as male or female (unsexed not included)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,sex=c("M","F"))

# extracts info for cod (081) 
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,species=81)

# extracts info for cod ("COD") 
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,"cod") or
get_length_age(channel,"co") or
get_length_age(channel,"COD")

}

}
\seealso{
\code{\link{connect_to_database}}, \code{\link{get_gears}}, \code{\link{get_species}}, \code{\link{get_areas}}, \code{\link{get_locations}}
, \code{\link{get_vessels}}, \code{\link{get_ports}}, \code{\link{get_anything_sql}}
}
