% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gears.R
\name{get_gears}
\alias{get_gears}
\title{Extract fishing GEAR types from CFDBS}
\usage{
get_gears(channel, gears = "all")
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{gears}{specific gear code or set of codes. Either numeric or character vector. Defaults to "all" gears.
Numeric codes are converted to VARCHAR2(2 BYTE) when creating the sql statement. Character codes are short character strings referencing GEARNM field.}
}
\value{
A list is returned:

  \item{gears}{containing the result of the executed \code{sqlStatement}}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}

If no \code{sqlStatement} is provided the default sql statement "\code{select * from cfdbs.gear}" is used
}
\description{
Extract a list of gear types in the NEFSC "GEAR" supporting table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations
}

\examples{
\dontrun{
# extracts gear data from cfdbs.gear table based on the default \code{sqlStatement}
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_gears(channel)

# extracts info based on gear types (5,6) (numeric)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_gears(channel,gears=c(5,6))

# extracts info based on gear types (5,6) (character)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_gears(channel,gears=c("05","06"))

# extracts info for "Seines"
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_gears(channel,"seines") or

}

}
\seealso{
\code{\link{connect_to_database}}

Other get functions: 
\code{\link{get_age_length}()},
\code{\link{get_anything_sql}()},
\code{\link{get_areas}()},
\code{\link{get_landings_length}()},
\code{\link{get_landings_palmer}()},
\code{\link{get_landings}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species_itis}()},
\code{\link{get_species}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
