% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landings_length.R
\name{get_landings_length}
\alias{get_landings_length}
\title{Extract landings data by length from ADIOS! (help from CFDBS)}
\usage{
get_landings_length(
  channel,
  area = "all",
  gear = "all",
  year = 1994,
  species = "all",
  species_itis = FALSE
)
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{species}{Numeric vector. Species codes (nespp3 or species_itis)}

\item{species_itis}{Boolean. Indicating if species values are nespp3 (FALSE) or species_itis (TRUE).Default = FALSE}

\item{areas}{Numeric vector. Statistical areas (area)}

\item{gears}{Numeric vector. Gear types (negear2)}

\item{years}{Numeric vector. containing the years to search over}
}
\value{
A list is returned:

  \code{data}      containing the result of the data pull ( year, month, day, permit, qtr, negear, negear2, nespp3, nespp4, species_itis, market_code, area, spplndlb, length, numlen, numsamp, sex  )

  \code{sql}  the resulting sql statement based on user input

  \code{colNames} a vector of the table's column names
}
\description{
Extract landings information by length (\code{len},\code{numlen},\code{spplndlb}) in cm/lbs from ADIOS! (mv_cf_len). This function requires no knowledge of sql.
For direct sql usage please see \code{\link{get_anything_sql}} function
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations
}

\section{Warning}{

You will need to obtain read only privilages to access ADIOS! 
The ADIOS landings by length database (MV_CF_LEN) contains 1.3 million records. If you try to pull the entire database you may suffer from memory issues.
}

\examples{
\dontrun{

# extracts landings by length for cod (081) in area 500 using  gear types (1,2,10)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")

get_landings_length(channel,area=c(500,600),gear=c(1,2,10),year=c(1999),species=81)
}

}
\seealso{
\code{\link{connect_to_database}}

Other get functions: 
\code{\link{get_age_length}()},
\code{\link{get_anything_sql}()},
\code{\link{get_areas}()},
\code{\link{get_gears}()},
\code{\link{get_landings}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species_itis}()},
\code{\link{get_species}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
