% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ports.R
\name{get_ports}
\alias{get_ports}
\title{Extract PORT location information from CFDBS}
\usage{
get_ports(channel, ports = "all")
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{ports}{a specific port code or set of codes. Either numeric or character vector. Defaults to "all" ports.
Numeric codes are converted to VARCHAR2(6 BYTE) when creating the sql statement. Character codes are short character strings referencing PORTNM field.}
}
\value{
A list is returned:

  \item{species}{containing the result of the executed \code{sqlStatement}}

   \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}

If no \code{sqlStatement} is provided the default sql statement "\code{select * from cfdbs.port}" is used
}
\description{
Extract a list of port names, and location info for vessel landings from the NEFSC "Port" supporting table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations
}

\examples{
\dontrun{
# extracts complete port table based on default sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_ports(channel)

# extracts port details based on ports 224309 , 224409 (numeric)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_ports(channel,ports = c(224309,224409))

# extracts port details based on ports 224309 , 224409 (character)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_ports(channel,ports = c("224309","224409"))

# extracts port details based on port name (character)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_ports(channel,ports = "Fairfield") or 
get_ports(channel,ports = "Fair")

}

}
\seealso{
\code{\link{connect_to_database}}, , \code{\link{get_gears}}, \code{\link{get_species}}, \code{\link{get_areas}}, \code{\link{get_locations}}
, \code{\link{get_vessels}}, \code{\link{get_ports}}
}
