% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Extract SPECIES information from CFDBS}
\usage{
get_species(channel, species = "all")
}
\arguments{
\item{channel}{an RODBC object (see \code{\link{connect_to_database}})}

\item{species}{a specific species code or set of codes. Either numeric or character vector. Defaults to "all" species.
Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement.}
}
\value{
A list is returned:

  \code{species} containing the result of the executed \code{$sql} statement

 \code{sql} containing the sql call

  \code{colNames} a vector of the table's column names

The default sql statement "\code{select * from cfdbs.cfspp}" is used
}
\description{
Extract a list of speices names, code, market category, etc from the NEFSC cfspp table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations. 
Note: species codes (nespp3) are stored in the database as VARCHAR2(3 BYTE)
}

\examples{
\dontrun{
# extracts complete species table based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species(channel)

# extracts info for cod (081) 
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species(channel,species=81)

# extracts info for cod (081)  and bluefish (023)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select * from cfdbs.cfspp"
get_species(channel,species= c("081","023"))



}

}
\seealso{
\code{\link{connect_to_database}}, \code{\link{get_gears}}, \code{\link{get_species}}, \code{\link{get_areas}}, \code{\link{get_locations}}
, \code{\link{get_vessels}}, \code{\link{get_ports}}, \code{\link{get_anything_sql}}
}
