% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessels.R
\name{get_vessels}
\alias{get_vessels}
\title{Extract VESSEL information from CFDBS}
\usage{
get_vessels(channel, sqlStatement = "select * from cfdbs.mstrvess",
  where = NULL)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{sqlStatement}{an sql statement (optional)}

\item{where}{text string appending where clause to sql}
}
\value{
A list is returned:

  \item{vessels}{containing the result of the executed \code{sqlStatement} }

  \item{sql}{containing the \code{sqlStatement} itself}

 \item{colNames}{a vector of the table's column names}

If no \code{sqlStatement} is provided the default sql statement "\code{select * from cfdbs.mstrvess}" is used
}
\description{
Extract a list of vessell ID's, tonnage, crew size, home port, etc from the NEFSC "Mstrvess" supporting table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations
}

\examples{
\dontrun{
# extracts complete vessel table based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_vessels(channel)

# extracts vessel ID, crew size, and home port on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select vessel, crew, homeport from cfdbs.mstrvess;"
get_vessels(channel,sqlStatement)

# extracts all vessel info for vessels lengths < 50ft based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select * from cfdbs.mstrvess where vesslen <29;"
get_vessels(channel,sqlStatement)
}

}
\seealso{
\code{\link{connect_to_database}}, \code{\link{get_gears}}, \code{\link{get_species}}, \code{\link{get_areas}}, \code{\link{get_locations}}
, \code{\link{get_vessels}}, \code{\link{get_ports}}
}
