% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas.R
\name{get_areas}
\alias{get_areas}
\title{Extract AREA information from CFDBS}
\usage{
get_areas(channel, areas = "all")
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{areas}{a specific area code or set of codes. Either numeric or character vector. Defaults to "all" areas
Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement. Character codes are short character strings to reference the AREANM field.}
}
\value{
A list is returned:

   \item{areas}{containing the result of the executed \code{sqlStatement}}

   \item{sql}{containing the sql call}

   \item{colNames}{ a vector of the table's column names}

The default sql statement "\code{select * from cfdbs.area}" is used
}
\description{
Extract a list of statistical areas, region, NAFO codes, etc from the NEFSC "Area" supporting table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations
}

\examples{
\dontrun{
# extracts complete area table based on default sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_areas(channel)

# extracts a subset of area data based on selected areas 100,500 (numeric)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_areas(channel,areas=c(100,500))

# extracts a subset of area data based on selected areas 100,500 (character)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_areas(channel,areas=c("100","500"))

# extracts a subset of area data based on areanm's containing "GG" (Androscoggin River etc)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_areas(channel,"GG")

}

}
\seealso{
\code{\link{connect_to_database}}

Other get functions: 
\code{\link{get_age_length}()},
\code{\link{get_anything_sql}()},
\code{\link{get_gears}()},
\code{\link{get_landings_length}()},
\code{\link{get_landings}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species_itis}()},
\code{\link{get_species}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
