% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Extract SPECIES information from CFDBS}
\usage{
get_species(channel, species = "all")
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{species}{a specific species code or set of codes. Either numeric or character vector. (NESPP3 codes)
Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement.
A Species common name can also be supplied. The character string is used to pull from SPPNM field. Defaults to "all" species.}
}
\value{
A list is returned:

  \item{species}{containing the result of the executed \code{$sql} statement}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}

The default sql statement "\code{select * from cfdbs.cfspp}" is used
}
\description{
Extract a list of speices names, code, market category, etc from the NEFSC cfspp table
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations. 
Note: species codes (nespp3) are stored in the database as VARCHAR2(3 BYTE)
}

\examples{
\dontrun{
# extracts complete species table based on custom sql statement
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species(channel)

# extracts info for cod (081) 
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species(channel,species=81)

# extracts info for cod ("COD") 
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_species(channel,"cod") or
get_species(channel,"co") or
get_species(channel,"COD")


# extracts info for cod (081)  and bluefish (023)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
sqlStatement <- "select * from cfdbs.cfspp"
get_species(channel,species= c("081","023"))
}

}
\seealso{
\code{\link{connect_to_database}}

Other get functions: 
\code{\link{get_age_length}()},
\code{\link{get_anything_sql}()},
\code{\link{get_areas}()},
\code{\link{get_gears}()},
\code{\link{get_landings_length}()},
\code{\link{get_landings_palmer}()},
\code{\link{get_landings}()},
\code{\link{get_locations}()},
\code{\link{get_ports}()},
\code{\link{get_species_itis}()},
\code{\link{get_vessels}()}
}
\concept{get functions}
